Support and use cufft extensible plan API
This change adds support for CUFFT's "extensible plan" API, in which
a handle is first created with cufftCreate and the actual plan made
later by cufftMakePlan*. This API is also used in the high-level Plan
object, since it allows parameters like compatibility mode to be set
before creating the plan (recommended by cufft docs).

Finally, support is added for caller-managed work area allocation. At
the low level, cufftSetAutoAllocation and cufftSetWorkArea are added. At
the high level, Plan() takes an auto_allocate argument, and adds a
Plan.set_work_area method.

This does not support the multi-GPU API, but it will be easier to add in
future because it is built on the extensible API. There are fundamental
issues with adding support, because for some pointer arguments there is
no way to know how many elements are being pointed to.